+++ src/System.IO.Pipelines/System.IO.Pipelines.csproj
