+++ src/System.IO.Pipelines/System/IO/Pipelines/ReadableBuffer.cs
@@ -20,7 +20,7 @@ namespace System.IO.Pipelines
         /// <summary>
         /// Length of the <see cref="ReadableBuffer"/> in bytes.
         /// </summary>
-        public long Length => GetLength(BufferStart, BufferEnd);
+        public long Length => ReadCursorOperations.GetLength(BufferStart, BufferEnd);
 
         /// <summary>
         /// Determines if the <see cref="ReadableBuffer"/> is empty.
@@ -36,7 +36,7 @@ namespace System.IO.Pipelines
         {
             get
             {
-                TryGetBuffer(BufferStart, BufferEnd, out var first, out _);
+                ReadCursorOperations.TryGetBuffer(BufferStart, BufferEnd, out var first, out _);
                 return first;
             }
         }
@@ -90,8 +90,8 @@ namespace System.IO.Pipelines
         /// <param name="length">The length of the slice</param>
         public ReadableBuffer Slice(long start, long length)
         {
-            var begin = Seek(BufferStart, BufferEnd, start, false);
-            var end = Seek(begin, BufferEnd, length, false);
+            var begin = ReadCursorOperations.Seek(BufferStart, BufferEnd, start, false);
+            var end = ReadCursorOperations.Seek(begin, BufferEnd, length, false);
             return new ReadableBuffer(begin, end);
         }
 
@@ -102,8 +102,8 @@ namespace System.IO.Pipelines
         /// <param name="end">The end (inclusive) of the slice</param>
         public ReadableBuffer Slice(long start, ReadCursor end)
         {
-            BoundsCheck(BufferEnd, end);
-            var begin = Seek(BufferStart, end, start);
+            ReadCursorOperations.BoundsCheck(BufferEnd, end);
+            var begin = ReadCursorOperations.Seek(BufferStart, end, start);
             return new ReadableBuffer(begin, end);
         }
 
@@ -114,8 +114,8 @@ namespace System.IO.Pipelines
         /// <param name="end">The ending (inclusive) <see cref="ReadCursor"/> of the slice</param>
         public ReadableBuffer Slice(ReadCursor start, ReadCursor end)
         {
-            BoundsCheck(BufferEnd, end);
-            BoundsCheck(end, start);
+            ReadCursorOperations.BoundsCheck(BufferEnd, end);
+            ReadCursorOperations.BoundsCheck(end, start);
 
             return new ReadableBuffer(start, end);
         }
@@ -127,9 +127,9 @@ namespace System.IO.Pipelines
         /// <param name="length">The length of the slice</param>
         public ReadableBuffer Slice(ReadCursor start, long length)
         {
-            BoundsCheck(BufferEnd, start);
+            ReadCursorOperations.BoundsCheck(BufferEnd, start);
 
-            var end = Seek(start, BufferEnd, length, false);
+            var end = ReadCursorOperations.Seek(start, BufferEnd, length, false);
 
             return new ReadableBuffer(start, end);
         }
@@ -140,7 +140,7 @@ namespace System.IO.Pipelines
         /// <param name="start">The starting (inclusive) <see cref="ReadCursor"/> at which to begin this slice.</param>
         public ReadableBuffer Slice(ReadCursor start)
         {
-            BoundsCheck(BufferEnd, start);
+            ReadCursorOperations.BoundsCheck(BufferEnd, start);
 
             return new ReadableBuffer(start, BufferEnd);
         }
@@ -153,7 +153,7 @@ namespace System.IO.Pipelines
         {
             if (start == 0) return this;
 
-            var begin = Seek(BufferStart, BufferEnd, start, false);
+            var begin = ReadCursorOperations.Seek(BufferStart, BufferEnd, start, false);
             return new ReadableBuffer(begin, BufferEnd);
         }
 
@@ -213,7 +213,7 @@ namespace System.IO.Pipelines
             {
                 throw new ArgumentOutOfRangeException(nameof(count));
             }
-            return Seek(cursor, BufferEnd, count, false);
+            return ReadCursorOperations.Seek(cursor, BufferEnd, count, false);
         }
 
         /// <summary>
@@ -263,204 +263,6 @@ namespace System.IO.Pipelines
             }
 
             return new ReadableBuffer(data, offset, length);
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        internal static long GetLength(ReadCursor begin, ReadCursor end)
-        {
-            if (begin.IsDefault)
-            {
-                return 0;
-            }
-
-            var segment = begin.Segment;
-            switch (segment)
-            {
-                case IMemoryList<byte> bufferSegment:
-                    return GetLength(bufferSegment, begin.Index, end.Get<IMemoryList<byte>>(), end.Index);
-                case byte[] _:
-                case OwnedMemory<byte> _:
-                    return end.Index - begin.Index;
-            }
-
-            PipelinesThrowHelper.ThrowInvalidOperationException(ExceptionResource.UnexpectedSegmentType);
-            return default;
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        internal static long GetLength(
-            IMemoryList<byte> start,
-            int startIndex,
-            IMemoryList<byte> endSegment,
-            int endIndex)
-        {
-            if (start == endSegment)
-            {
-                return endIndex - startIndex;
-            }
-
-            return (endSegment.VirtualIndex - start.Next.VirtualIndex)
-                   + (start.Memory.Length - startIndex)
-                   + endIndex;
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        internal static ReadCursor Seek(ReadCursor begin, ReadCursor end, long bytes, bool checkEndReachable = true)
-        {
-            ReadCursor cursor;
-            if (begin.Segment == end.Segment && end.Index - begin.Index >= bytes)
-            {
-                // end.Index >= bytes + Index and end.Index is int
-                cursor = new ReadCursor(begin.Segment, begin.Index + (int)bytes);
-            }
-            else
-            {
-                cursor = SeekMultiSegment(begin, end, bytes, checkEndReachable);
-            }
-
-            return cursor;
-        }
-
-        [MethodImpl(MethodImplOptions.NoInlining)]
-        private static ReadCursor SeekMultiSegment(ReadCursor begin, ReadCursor end, long bytes,  bool checkEndReachable)
-        {
-            ReadCursor result = default;
-            var foundResult = false;
-            var current = begin;
-            while (TryGetBuffer(begin, end, out var memory, out begin))
-            {
-                // We need to loop up until the end to make sure start and end are connected
-                // if end is not trusted
-                if (!foundResult)
-                {
-                    // We would prefer to put cursor in the beginning of next segment
-                    // then past the end of previous one, but only if next exists
-
-                    if (memory.Length > bytes ||
-                       (memory.Length == bytes && begin.IsDefault))
-                    {
-                        result = new ReadCursor(current.Segment, current.Index + (int)bytes);
-                        foundResult = true;
-                        if (!checkEndReachable)
-                        {
-                            break;
-                        }
-                    }
-
-                    bytes -= memory.Length;
-                }
-                current = begin;
-            }
-
-            if (!foundResult)
-            {
-                PipelinesThrowHelper.ThrowCursorOutOfBoundsException();
-            }
-
-            return result;
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        internal static void BoundsCheck(ReadCursor end, ReadCursor newCursor)
-        {
-            switch (end.Segment)
-            {
-                case byte[] _ :
-                case OwnedMemory<byte> _ :
-                    if (newCursor.Index > end.Index)
-                    {
-                        PipelinesThrowHelper.ThrowCursorOutOfBoundsException();
-                    }
-                    return;
-                case IMemoryList<byte> memoryList:
-                    if (!GreaterOrEqual(memoryList, end.Index, newCursor.Get<IMemoryList<byte>>(), newCursor.Index))
-                    {
-                        PipelinesThrowHelper.ThrowCursorOutOfBoundsException();
-                    }
-                    return;
-                default:
-                    PipelinesThrowHelper.ThrowCursorOutOfBoundsException();
-                    return;
-            }
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        internal static bool TryGetBuffer(ReadCursor begin, ReadCursor end, out ReadOnlyMemory<byte> data, out ReadCursor next)
-        {
-            var segment = begin.Segment;
-
-            switch (segment)
-            {
-                case null:
-                    data = default;
-                    next = default;
-                    return false;
-
-                case IMemoryList<byte> bufferSegment:
-                    var startIndex = begin.Index;
-                    var endIndex = bufferSegment.Memory.Length;
-
-                    if (segment == end.Segment)
-                    {
-                        endIndex = end.Index;
-                        next = default;
-                    }
-                    else
-                    {
-                        var nextSegment = bufferSegment.Next;
-                        if (nextSegment == null)
-                        {
-                            if (end.Segment != null)
-                            {
-                                PipelinesThrowHelper.ThrowInvalidOperationException(ExceptionResource.EndCursorNotReached);
-                            }
-
-                            next = default;
-                        }
-                        else
-                        {
-                            next = new ReadCursor(nextSegment, 0);
-                        }
-                    }
-
-                    data = bufferSegment.Memory.Slice(startIndex, endIndex - startIndex);
-
-                    return true;
-
-
-                case OwnedMemory<byte> ownedMemory:
-                    data = ownedMemory.Memory.Slice(begin.Index, end.Index - begin.Index);
-
-                    if (segment != end.Segment)
-                    {
-                         PipelinesThrowHelper.ThrowInvalidOperationException(ExceptionResource.EndCursorNotReached);
-                    }
-
-                    next = default;
-                    return true;
-
-                case byte[] array:
-                    data = new Memory<byte>(array, begin.Index, end.Index - begin.Index);
-
-                    if (segment != end.Segment)
-                    {
-                        PipelinesThrowHelper.ThrowInvalidOperationException(ExceptionResource.EndCursorNotReached);
-                    }
-                    next = default;
-                    return true;
-            }
-
-            PipelinesThrowHelper.ThrowNotSupportedException();
-            next = default;
-            return false;
-        }
-
-        internal static bool GreaterOrEqual(IMemoryList<byte> start, int startIndex, IMemoryList<byte> end, int endIndex)
-        {
-            // other.Segment.RunningLength + other.Index  - other.Segment.Start <= Segment.RunningLength + Index- Segment.Start
-            // fliped to avoid overflows
-
-            return end.VirtualIndex - startIndex <= start.VirtualIndex - endIndex;
         }
     }
 }
